Milestone 1/requirements.txt.txt
Flask==3.1.2
python-dotenv==1.2.1
Authlib==1.6.6

# Face authentication
face-recognition==1.3.0
opencv-python==4.12.0.88
numpy==2.2.6
dlib-bin==19.24.6

# Gmail API
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0

# Voice input
SpeechRecognition
pyaudio
openai-whisper==20250625
